% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alimena_square}
\alias{alimena_square}
\title{Generate an alimena square}
\usage{
alimena_square(labels)
}
\arguments{
\item{labels}{The labels to be filled in as values in the latin square.}
}
\description{
Generate an alimena square
}
