% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stim_lists.R
\name{compose_data}
\alias{compose_data}
\title{Compose response data from fixed and random effects}
\usage{
compose_data(design_args, fixed = NULL, subj_rmx = NULL,
  item_rmx = NULL, verbose = FALSE, contr_type = "contr.dev")
}
\arguments{
\item{design_args}{List containing information about the experiment
design; see \code{\link{stim_lists}}}

\item{fixed}{vector of fixed effects}

\item{subj_rmx}{matrix of by-subject random effects}

\item{item_rmx}{matrix of by-item random effects}

\item{verbose}{give debugging info (default = \code{FALSE})}

\item{contr_type}{contrast type (default "contr.dev"); see ?contrasts}
}
\value{
a data frame containing response variable \code{Y}, the
linear sum of all fixed and random effects.
}
\description{
Compose response data from fixed and random effects
}
\details{
This will add together all of the fixed and random effects
according to the linear model specified by the design.  Note,
however, that it does not add in any residual noise; for that, use
the function \code{\link{sim_norm}}.
}
