% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{contr.dev}
\alias{contr.dev}
\title{Deviation-Coded Contrast Matrices}
\usage{
contr.dev(n, base = 1, contrasts = TRUE)
}
\arguments{
\item{n}{a vector of levels for a factor, or the number of levels.}

\item{base}{an integer specifying which group is considered the
baseline group. Ignored if 'contrasts' is \code{FALSE}.}

\item{contrasts}{a logical indicating whether contrasts should be computed.}
}
\description{
Return a matrix of deviation-coded contrasts.
}

