% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{design_formula}
\alias{design_formula}
\title{Get the GLM formula for a factorially-designed experiment}
\usage{
design_formula(design_args, n_subj = NULL, dv_name = NULL, lme4_format = TRUE)
}
\arguments{
\item{design_args}{A list with experimental design information (see \code{link{stim_lists}})}

\item{n_subj}{Number of subjects}

\item{dv_name}{Name of dependent variable; \code{NULL} (default) for one-sided formula}

\item{lme4_format}{Do you want the results combined as the model formula for a \code{lme4} model? (default \code{TRUE})}
}
\value{
A formula, character string, or list, depending
}
\description{
Get the formula corresponding to the general linear model for a
factorially designed experiment, with maximal random effects.
}
