% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{digram_square_even}
\alias{digram_square_even}
\title{Generate a digram-balanced latin square for even n}
\usage{
digram_square_even(labels)
}
\arguments{
\item{labels}{The labels to be filled in as values in the latin square.}
}
\description{
Generate a digram-balanced latin square for even n
}
