% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{digram_square_odd}
\alias{digram_square_odd}
\title{Generate a digram-balanced latin squares for odd n}
\usage{
digram_square_odd(labels)
}
\arguments{
\item{labels}{The labels to be filled in as values in the latin square.}
}
\description{
Generate a digram-balanced latin squares for odd n
}
