% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{fac_counts}
\alias{fac_counts}
\title{Calculate marginal and cell counts for factorially-designed data}
\usage{
fac_counts(iv_names, dat, unit_names = c("subj_id", "item_id"))
}
\arguments{
\item{iv_names}{Names of independent variables in data.frame given by \code{dat}.}

\item{dat}{A data frame}

\item{unit_names}{Names of the fields containing sampling units (subjects, items)}
}
\value{
a list, the elements of which have the marginal/cell counts for each factor in the design
}
\description{
Calculate marginal and cell counts for factorially-designed data
}

