% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generic_square}
\alias{generic_square}
\title{Generate a generic latin square}
\usage{
generic_square(labels)
}
\arguments{
\item{labels}{The labels to be filled in as values in the latin square.}
}
\description{
Generate a generic latin square
}
