% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{prmd_factor}
\alias{prmd_factor}
\title{Generate pseudorandom minimum distance sequence for factors}
\usage{
prmd_factor(n, levels)
}
\description{
Generate pseudorandom minimum distance sequence for factors
}
