% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_norm}
\alias{sim_norm}
\title{Sample data from population with normal error variance}
\usage{
sim_norm(design_args, n_subj, params)
}
\arguments{
\item{design_args}{experiment design (see
\code{\link{stim_lists}}); must contain sub-element \code{n_item}
defining the number of items.}

\item{n_subj}{number of subjects}

\item{params}{list with parameters defining population
(\code{fixed}, \code{subj_rfx}, \code{item_rfx}), and
\code{err_var}; normally generated by a call to
\code{\link{gen_pop}}.}
}
\value{
A data frame with simulated data
}
\description{
Sample data from population with normal error variance
}
\examples{
design_args <- list(ivs = c(A = 2, B = 3), n_item = 18)
pop_params <- gen_pop(design_args, 12)

dat <- sim_norm(design_args, 12, pop_params)
}
\seealso{
\code{\link{gen_pop}}
}

