% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stim_lists.R
\name{stim_lists}
\alias{stim_lists}
\title{Generate stimulus presentation lists}
\usage{
stim_lists(design_args, as_one = TRUE)
}
\arguments{
\item{design_args}{A list describing the experimental design, which
must have an element \code{ivs}, giving a named list of independent
variables, with each list element a vector giving the levels of
that IV, or a single number stating the number of desired levels.
If any IVs are administered between-subjects or between-items,
these should be named in list elements \code{between_subj} and
\code{between_item} respectively.  The argument \code{design_args}
also can optionally include the following two elements:
\code{n_item}, the desired number of stimulus items, which if
unspecified, will result in lists with the minimum possible number
of items; and \code{n_rep}, the number of repetitions of each
stimulus item for each participant (default 1).}

\item{as_one}{boolean (default \code{TRUE}) specifying whether the
presentation lists are to be returned as a single data frame or as
elements in a list object}
}
\value{
a single \code{data.frame} (default) with each list
identified by \code{list_id} or a \code{list} of dataframes,
depending on the value of \code{as_one}
}
\description{
Generates counterbalanced presentation lists for factorially
designed experiments involving stimulus presentation
}
\examples{
stim_lists(list(ivs = c(A = 2, B = 2))) # 2x2 within-subjects within-item

stim_lists(list(ivs = c(A = 2, B = 2, n_item = 16))) # same but w/more items

stim_lists(list(ivs = c(A = 2, B = 2, n_item = 16, n_rep = 3)))

# mixed by subject, fully within by item
stim_lists(list(ivs = list(group = c("adult", "child"),
                           task = c("easy", "hard")),
                between_subj = "group",
                n_item = 12))

# mixed by subject, mixed by item
stim_lists(list(ivs = c(A = 2, B = 2),
           between_subj = "A",
           between_item = "B"))
}

