% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{term_names}
\alias{term_names}
\title{Get names for predictors in factorial design}
\usage{
term_names(design_args, design_formula = NULL, contr_type = "contr.dev")
}
\arguments{
\item{design_args}{A list with experimental design information (see \code{link{stim_lists}})}

\item{design_formula}{A formula (default NULL, constructs from \code{design_args})}

\item{contr_type}{Name of formula for generating contrasts (default "contr.dev")}
}
\value{
A character vector with names of all the terms
}
\description{
Get the names for the numerical predictors corresponding to all
main effects and interactions of categorical IVs in a factorial
design.
}

