% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stim_lists.R
\name{trial_lists}
\alias{trial_lists}
\title{Generate trial lists from a stimulus presentation lists}
\usage{
trial_lists(design_args, subjects = NULL, seq_assign = FALSE)
}
\arguments{
\item{design_args}{Stimulus presentation lists (see \code{\link{stim_lists}}).}

\item{subjects}{One of the following three: (1) an integer
specifying the desired number of subjects (must be a multiple of
number of stimulus lists); (2) a data frame with assignment
information (must include a column \code{subj_id}); or (3)
\code{NULL}, in which case there will be one subject per list.}

\item{seq_assign}{If TRUE, assignment of subjects to lists will
be sequential rather than random (default is FALSE)}
}
\value{
A data frame containing all trial information.
}
\description{
Merge stimulus presentation lists with subject data to create a
trial list.
}

