% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{with_dev_pred}
\alias{with_dev_pred}
\title{Generate numerical deviation-coded predictors}
\usage{
with_dev_pred(dat, iv_names = NULL)
}
\arguments{
\item{dat}{A data frame with columns containing factors to be converted.}

\item{iv_names}{Names of the variables to be converted.}
}
\value{
A data frame including additional deviation coded predictors.
}
\description{
Add deviation-coded predictors to data frame.
}
\examples{
with_dev_pred(stim_lists(list(ivs = c(A = 3))), "A")
}
