% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_norm}
\alias{sim_norm}
\title{Sample data from population with normal error variance}
\usage{
sim_norm(mcr.data)
}
\arguments{
\item{mcr.data}{list with parameters defining population
(\code{fixed}, \code{subj_rfx}, \code{item_rfx}).  Must also
contain element \code{design_args}, defining the experiment design,
and \code{design_args} must also contain the sub-element
\code{n_item} defining the number of stimulus items.  Finally,
\code{mcr.data} must contain element \code{n_subj}, defining the
number of subjects to be sampled.}
}
\value{
A data frame
}
\description{
Sample data from population with normal error variance
}
\examples{
design_args <- list(ivs = c(A = 2, B = 3), n_item = 18)
mcr.data <- gen_pop(design_args, n_subj = 12)
mcr.data[["n_subj"]] <- 12
mcr.data[["design_args"]] <- design_args
mcr.data[["err_var"]] <- 3
dat <- sim_norm(mcr.data)
}
\seealso{
\code{\link{gen_pop}}
}

